Renaming, rewording and removing unused logger. Just a few small things.
